Added steps to detail VC with appropriate resizing for varying screen sizes
