+++ How-To/How-To/Helpers/UIView+Layout.swift
@@ -272,3 +272,12 @@ extension UIView {
         self.layer.shadowPath = path.cgPath
 }
 }
+
+
+extension UILabel {
+    convenience init(text: String, font: UIFont) {
+        self.init(frame: .zero)
+        self.text = text
+        self.font = font
+    }
+}
